update inprocess kernel to new layout
including Qt mixins
