+++ IPython/frontend/qt/inprocess.py
@@ -2,12 +2,14 @@
 """
 
 # Local imports.
-from IPython.kernel.inprocess.kernelmanager import \
-    InProcessShellChannel, InProcessIOPubChannel, InProcessStdInChannel, \
-    InProcessHBChannel, InProcessKernelManager
+from IPython.kernel.inprocess import (
+    InProcessShellChannel, InProcessIOPubChannel, InProcessStdInChannel,
+    InProcessHBChannel, InProcessKernelClient
+)
+
 from IPython.utils.traitlets import Type
-from base_kernelmanager import QtShellChannelMixin, QtIOPubChannelMixin, \
-    QtStdInChannelMixin, QtHBChannelMixin, QtKernelManagerMixin
+from kernel_mixins import QtShellChannelMixin, QtIOPubChannelMixin, \
+    QtStdInChannelMixin, QtHBChannelMixin, QtKernelClientMixin
 
 
 class QtInProcessShellChannel(QtShellChannelMixin, InProcessShellChannel):
@@ -22,8 +24,7 @@ class QtInProcessStdInChannel(QtStdInChannelMixin, InProcessStdInChannel):
 class QtInProcessHBChannel(QtHBChannelMixin, InProcessHBChannel):
     pass
 
-
-class QtInProcessKernelManager(QtKernelManagerMixin, InProcessKernelManager):
+class QtInProcessKernelClient(QtKernelClientMixin, InProcessKernelClient):
     """ An in-process KernelManager with signals and slots.
     """
