+++ IPython/kernel/inprocess/__init__.py
@@ -1,0 +1,8 @@
+from .channels import (
+    InProcessShellChannel,
+    InProcessIOPubChannel,
+    InProcessStdInChannel,
+    InProcessHBChannel,
+)
+from .client import InProcessKernelClient
+from .manager import InProcessKernelManager
